+++ src/main/java/ru/yandex/practicum/filmorate/model/Review.java
@@ -10,15 +10,15 @@ import javax.validation.constraints.NotNull;
 @NoArgsConstructor
 @Builder
 public class Review {
-    private Integer reviewId;
+    private Long reviewId;
     @NotBlank(message = "Review content must be completed")
     private String content;
     @NotNull(message = "Review type must be positive(true) or negative(false)")
     private Boolean isPositive;
     @NotNull(message = "Review userId cannot be empty")
-    private Integer userId;
+    private Long userId;
     @NotNull(message = "Review filmId cannot be empty")
-    private Integer filmId;
+    private Long filmId;
     @Setter(AccessLevel.NONE)
     private int useful;
 }
