+++ src/main/java/ru/yandex/practicum/filmorate/storage/dao/review/impl/ReviewDaoImpl.java
@@ -30,13 +30,13 @@ public class ReviewDaoImpl implements ReviewDao {
             PreparedStatement ps = connection.prepareStatement(sqlQuery, new String[]{"review_id"});
             ps.setString(1, review.getContent());
             ps.setBoolean(2, review.getIsPositive());
-            ps.setInt(3, review.getUserId());
-            ps.setInt(4, review.getFilmId());
+            ps.setLong(3, review.getUserId());
+            ps.setLong(4, review.getFilmId());
             ps.setInt(5, 0);
             return ps;
         }, keyHolder);
 
-        review.setReviewId(Objects.requireNonNull(keyHolder.getKey()).intValue());
+        review.setReviewId(Objects.requireNonNull(keyHolder.getKey()).longValue());
         return review;
     }
 
@@ -51,12 +51,12 @@ public class ReviewDaoImpl implements ReviewDao {
     }
 
     @Override
-    public Review findById(Integer id) {
+    public Review findById(Long id) {
         return jdbcTemplate.queryForObject("SELECT * FROM reviews WHERE review_id = ?", this::mapRowToReview, id);
     }
 
     @Override
-    public List<Review> findAll(Integer filmId, Integer count) {
+    public List<Review> findAll(Long filmId, Integer count) {
         String sqlQuery;
 
         if (filmId != null) {
@@ -70,12 +70,12 @@ public class ReviewDaoImpl implements ReviewDao {
     }
 
     @Override
-    public void delete(Integer id) {
+    public void delete(Long id) {
         jdbcTemplate.update("DELETE FROM reviews WHERE review_id = ?", id);
     }
 
     @Override
-    public boolean checkExist(Integer id) throws NotFoundException {
+    public boolean checkExist(Long id) throws NotFoundException {
         String sqlQuery = "SELECT review_id FROM reviews WHERE review_id = ?";
         SqlRowSet rowSet = jdbcTemplate.queryForRowSet(sqlQuery, id);
         if (!rowSet.next()) {
@@ -86,11 +86,11 @@ public class ReviewDaoImpl implements ReviewDao {
 
     private Review mapRowToReview(ResultSet rs, int rowNum) throws SQLException {
         return Review.builder()
-                .reviewId(rs.getInt("review_id"))
+                .reviewId(rs.getLong("review_id"))
                 .content(rs.getString("content"))
                 .isPositive(rs.getBoolean("isPositive"))
-                .userId(rs.getInt("user_id"))
-                .filmId(rs.getInt("film_id"))
+                .userId(rs.getLong("user_id"))
+                .filmId(rs.getLong("film_id"))
                 .useful(rs.getInt("useful"))
                 .build();
     }
