+++ src/main/java/ru/yandex/practicum/filmorate/storage/dao/user/EventDao.java
@@ -5,7 +5,7 @@ import ru.yandex.practicum.filmorate.model.Event;
 import java.util.List;
 
 public interface EventDao {
-    void add(Integer userId, String eventType, String operation, Integer entityId);
+    void add(Long userId, String eventType, String operation, Long entityId);
 
-    List<Event> findUserFeed(Integer userId);
+    List<Event> findUserFeed(Long userId);
 }
